// admin/src/app/models/hotel-details.model.ts
export class Hotel_Details {
    Id: null;
    LocationID: number;
    CountryID: number;
    CountryLocationID: number;
    CityID: number;
    CityLocationID: number;
    HotelName: string;
    HomeTitle: string;
    HomeDescription: string;
    Address: string;
    TelephoneNumber: string;
    WebsiteUrl: string;
    LanguageValue: string;
    CreatedOn: null;
}